FEAT: Loan application process

+++ src/main/java/io/ezra/lending/entities/LoanApplication.java
@@ -5,6 +5,7 @@ import lombok.Getter;
 import lombok.Setter;
 
 import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.time.LocalDate;
 
 @Getter
@@ -18,7 +19,7 @@ public class LoanApplication {
 
     @Id
     @Column(name = "loan_reference_id", nullable = false)
-    private Integer loanReferenceId;
+    private BigInteger loanReferenceId;
 
     @Column(name = "customer_id")
     private Integer customerId;
